Interfaced WebSocket with Feed.
Had to make a container (I went with a stack, but should probably use a
queue) to accumulate messages as they weren't accumulated in the main
thread or some jazz. Also made it so you can click an entry and have it
move to the selected entry box and be removed from the feed list.