Changed Various Things
Made entry size smaller so they can move slower and there can be more on
the feed at once.
Changed message stack to a queue so the first messages received are the
first items added to the feed instead of the last. The problem with the
stack is that if loads of messages are received quickly its possible the
earliest items would never get added to the feed.
Made the entries be added under the feed box if the last entry in the
feed is above the bottom of the feed box, so entries can't appear out of
nowhere and are always coming from the bottom of the feed box.