More Window Changes + Browser Loading
Made it so windows have an order so they can be dragged in front of each
other.
Started the parsing of text from revision differences. Found the tag for
deletions and additions and looped through text to find it. Need to
extract it in readable format now.
Also gave shaders more suitable names and made 3D text have depth
testing.