Scrollbar, Scrolling, Parsing, Wrapping
Added a scrollbar prefab and script.
Made the browser text able to be scrolled. (Still needs tweaking, scroll
distance is odd...).
Made the scrollbar scale with the size of the data.
Parsed the data by removing html tags by looping through the text's
characters and looking for '<' and '>' tag start and end characters.
(Still needs more tags and stuff to be removed/resolved).
Added word wrapping. Since some links are too long for even one line,
and some of the annoying extra tags are too, i also split the line even
further by hyphenating it. (I should probably not do all the string
changing on the actual text, but on the word itself, so its faster and
doesnt assign so much memory every time).
Also windows can now be collapsed and expanded. Functionality for
minimisation is also there. The functions of the windows continues even
while minimised/collapsed so the feed for example still runs, just its
renderers are inactive.