Made cube positions be precalculated.
Moved the function to calculate cube positions out of the AddCube()
function and into PrecalculateCubePositions(). This is called once at
the start by the NodeMaster and calculates all local cube positions for
each allowed core size. This has removed practically all the lag since
there is no longer the need to run tonnes of calculations each time a
cube is added.