Implemented Flocking Enemies and Interception
Enemies flock as boids.
When the player is in sight (in a cone ahead of the enemy specified by
an angle and a distance) the enemy will shoot in the direction that is
predicted to cause an interception in the shortest time. Enemies spawn
from infected nodes. Problem right now with clumping enemies. Also,
enemies need a maximum distance from their owner node.