New Glow Shader + Lag Fix
Lag was present due to the parenting-unparenting-parenting etc. of cubes
and nodes to and from the nodemaster. Cubes are now never parented to
anything and are simply moved instead.  This means the nodes no longer
rotate so as to not require updating the cube positions every frame.
A new shader gives the cubes a kind of wireframe outline. The colour is
a bit messed up (albedo is multiplied by 100) but it means they can have
an outline and still be batched since its not multipass.
When spawning nodes and cubes, it is now made sure that theyre visible
for at least half a second before being deactivated. This removes the
possibility of any kind of lag being produced where they are first
rendered in-game. The delay could be reduced to just one frame, but this
way the loading screen doesn't jitter so much (yeah i'm prioritising the
loading screen over the time :P)